backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Image Processing
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.26.2

# AI/ML Libraries
torch==2.1.1
torchvision==0.16.1
easyocr==1.7.1

# LaMa Inpainting (simple-lama-inpainting)
simple-lama-inpainting==0.1.0

# Async Processing
celery[redis]==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.25.2